moved StackedMassBalance
